% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{dd_cluster_stations}
\alias{dd_cluster_stations}
\title{dd_cluster_stations}
\usage{
dd_cluster_stations(cl, city, stns, min_size = 3)
}
\arguments{
\item{cl}{Vector of cluster numbers obtained from \link{dd_cov_clusters}}

\item{city}{City for which clusters were obtained}

\item{stns}{A \pkg{tibble} of station data, including \code{stn_id},
\code{longitude}, and \code{latitude}, typically obtained from
\code{bikedata::bike_stations(city)}.}

\item{min_size}{Lower threshold above which to plot clusters}
}
\value{
A \pkg{tibble} of (\code{stn_id}, \code{longitude}, \code{latitude},
and \code{cl}), where the latter denotes the cluster number.
}
\description{
Convert the results of \link{dd_cov_clusters} to a \pkg{tibble} including
station IDs, longitudes, and latitudes.
}
