% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-models.R
\name{dd_compare_models}
\alias{dd_compare_models}
\title{compare_models}
\usage{
dd_compare_models(city = "nyc", from = TRUE, mi = FALSE, plot = FALSE)
}
\arguments{
\item{city}{City for which models are to be compared}

\item{from}{Plot covariance or MI values for trips \strong{from} each
station. If \code{FALSE}, plot equivalent values for trips \strong{to} each
station.}

\item{mi}{If \code{TRUE}, plot decay functions for mutual information,
otherwise plot covariances.}

\item{plot}{= TRUE allows visual inspection of individual station results}
}
\value{
a list of data frames, one for each model, containg data frame with
rows for each station and columns of [SS, k, b, AIC], where k is the width
parameter of the model, and b is the shape parameter (which does not apply to
power-law, Gaussian or Cauchy models).
}
\description{
Statistical comparisons of 6 different functional forms of distance decay:
power-law, Gaussian, exponential, Weibull, Cauchy, and Box-Cox
}
