% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-matrices.R
\name{dd_cov}
\alias{dd_cov}
\title{dd_cov}
\usage{
dd_cov(city)
}
\arguments{
\item{city}{City for which distance matrix is to be extracted}
}
\description{
Calculate station-to-station covariance matrix for a given city
}
\note{
The directory from which distance matrices are loaded can be retrieveed
with \link{dd_get_data_dir}, and set with \link{dd_set_data_dir}.
}
