% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusters.R
\name{dd_cov_clusters}
\alias{dd_cov_clusters}
\title{dd_cov_clusters}
\usage{
dd_cov_clusters(city = "nyc")
}
\arguments{
\item{city}{City for which models are to be compared}
}
\value{
nuffin
}
\description{
Cluster points which have covariances above values expected from a neutral
model.
}
