% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-models.R
\name{dd_decay_fns}
\alias{dd_decay_fns}
\title{dd_decay_fns}
\usage{
dd_decay_fns(city, from = TRUE)
}
\arguments{
\item{city}{City for which decay parameters are to be calculated}

\item{from}{Plot covariance or MI values for trips \strong{from} each
station. If \code{FALSE}, plot equivalent values for trips \strong{to} each
station.}

\item{mi}{If \code{TRUE}, plot decay functions for mutual information,
otherwise plot covariances.}
}
\value{
A \code{data.frame} of statistics for both straight-line distances
(\code{_str}) and OpenStreetMap network distances (\code{_osm}) for the
exponential decay function, exp (-(d / k) ^ b), of:
\itemize{
 \item \code{k}, the width of the exponential;
 \item \code{b}, the exponent; and
 \item \code{ss}, a standardised sum of sqaured residuals of the model fit
}
k-values quantifying the widths of exponential fits to the decay functions.
}
\description{
dd_decay_fns
}
