% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-models.R
\name{dd_fit_expmod}
\alias{dd_fit_expmod}
\title{dd_fit_expmod}
\usage{
dd_fit_expmod(mats, i, from = TRUE, plot = FALSE)
}
\arguments{
\item{mats}{List of distance and trip matrices for a particular city as
returned from \code{dd_get_tripdistmats()}.}

\item{i}{Number of station as row or column of matrices}

\item{from}{Analyse decay functions for trips \strong{from} each station. If
\code{FALSE}, analyse for trips \strong{to} each station.}

\item{plot}{If \code{TRUE}, plot decay function}
}
\value{
A \code{data.frame} of four values: \code{id}, the ID of the station;
\code{k}, the width parameter of the exponential decay; \code{b}, the value
of the exponent; and \code{ss}, the standardised sum of squared residuals.
}
\description{
Fit a generalised exponential decay model to data from one station
}
