% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matrices.R
\name{dd_get_tripmat}
\alias{dd_get_tripmat}
\title{dd_get_tripmat}
\usage{
dd_get_tripmat(city)
}
\arguments{
\item{city}{City for which trip matrix is to be extracted}
}
\description{
Load pre-calculated trip matrix for a given city
}
\note{
The directory from which trip matrices are loaded can be retrieveed
with \link{dd_get_data_dir}, and set with \link{dd_set_data_dir}.
}
