% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov-mi-matrices.R
\name{dd_mi}
\alias{dd_mi}
\title{dd_mi}
\usage{
dd_mi(city, lower = 0, upper = 1, osm = TRUE)
}
\arguments{
\item{city}{City for which mutual information matrix is to be extracted}

\item{lower}{Lower limit (0-1) for distance cutoff used to calculate
covariances (see details)}

\item{upper}{Upper limit (0-1) for distance cutoff used to calculate
covariances (see details)}

\item{osm}{If \code{FALSE}, return straight-line distances, otherwise street
network distances.}
}
\description{
Calculate station-to-station mutual information matrix for a given city.
Upper diagonal of matrix is between all trips \strong{from} i and j, so
\code{dd_cov()[i, j]} holds the covariance between all trips \strong{from}
station \code{i} to all other stations and \strong{from} station \code{j} to
all other stations. Analogously, the lower diagonal (\code{dd_cov()[j, i]})
holds covariances between all trips \strong{to} i and j.
}
\note{
The directory from which trip matrices are loaded can be retrieveed
with \link{dd_get_data_dir}, and set with \link{dd_set_data_dir}.
}
