% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fns.R
\name{dd_plot}
\alias{dd_plot}
\title{dd_plot}
\usage{
dd_plot(city, from = TRUE, mi = FALSE)
}
\arguments{
\item{city}{City for which distance decay functions are to be plotted. If
absent, all cities are plotted together.}

\item{from}{Plot covariance or MI values for trips \strong{from} each
station. If \code{FALSE}, plot equivalent values for trips \strong{to} each
station.}

\item{mi}{If \code{TRUE}, plot decay functions for mutual information,
otherwise plot covariances.}
}
\description{
Plot distance decay functions.
}
