% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fns.R
\name{dd_plot_all}
\alias{dd_plot_all}
\title{dd_plot_all}
\usage{
dd_plot_all(from = TRUE, mi = FALSE, smoother = FALSE)
}
\arguments{
\item{from}{Plot covariance or MI values for trips \strong{from} each
station. If \code{FALSE}, plot equivalent values for trips \strong{to} each
station.}

\item{mi}{If \code{TRUE}, plot decay functions for mutual information,
otherwise plot covariances.}

\item{smoother}{if \code{TRUE}, overlay a loess-smoothed line. (This will
generally take much longer to plot.)}
}
\description{
Plot distance decay functions for all cities
}
