% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-fns.R
\name{dd_plot_clusters}
\alias{dd_plot_clusters}
\title{dd_plot_clusters}
\usage{
dd_plot_clusters(cl, interactive = TRUE)
}
\arguments{
\item{cl}{Tibble of station data and cluster numbers obtained from
\link{dd_cluster_stations}}

\item{interactive}{If \code{FALSE}, produce a static \pkg{ggplot2} object,
otherwise an interactive map via \pkg{mapview}.}
}
\description{
Plot the results of the \link{dd_cov_clusters} function
}
