% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{dd_set_data_dir}
\alias{dd_set_data_dir}
\title{dd_set_data_dir}
\usage{
dd_set_data_dir(dir)
}
\arguments{
\item{dir}{Directory of pre-calculated distance matrices}
}
\description{
Set the directory of pre-calculated distance matrices
}
