% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distdecay-package.R
\docType{package}
\name{distdecay}
\alias{distdecay}
\alias{distdecay-package}
\title{Calcalate distance decay functions based on (dis)similarity matrices}
\description{
Input: One square matrix of (dis-)similarities between pairs of locations; one
squre matrix of corresponding distances.
}
\details{
Output: Distance decay functions for each location, quantified as either
covariance or mutual information coefficients.
}
\author{
Mark Padgham
}
