% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distmat.R
\name{distmat_g}
\alias{distmat_g}
\title{Calculate distance matrix between pairs of points using the google API}
\usage{
distmat_g(xy, dmat, g_units = "metric", g_mode = "bicycling")
}
\arguments{
\item{xy}{A two-column matrix of lon-lat coordinates}

\item{dmat}{If submitted, any missing values not returned in previous queries
are filled.}

\item{g_units}{The units of measurement requested from the google API}

\item{g_mode}{The mode of transport requested from the google API}
}
\description{
Calculate distance matrix between pairs of points using the google API
}
\note{
the google server frequently returns no values, so this produces a
matrix with lots of missing values.
}
\examples{
\dontrun{
test_fn ()
}
}
