% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distdecay-package.R
\docType{data}
\name{distmats}
\alias{distmats}
\title{distmats}
\description{
Station-to-station distance matrices between bicycle hire stations for all
cities of the \pkg{bikedata} package, calculated with the \code{bike_distmat}
function. These matrices have been reconciled against the corresponding
\link{tripmats} matrices with the function
\code{bikedata::bike_match_matrices}.
}
