% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distdecay-package.R
\docType{data}
\name{distmats_straight}
\alias{distmats_straight}
\title{distmats_straight}
\description{
Station-to-station distance matrices between bicycle hire stations for all
cities of the \pkg{bikedata} package, calculated as straight line distances.
These matrices have been reconciled against the corresponding \link{tripmats}
matrices with the function \code{bikedata::bike_match_matrices}.
}
