% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distdecay-package.R
\docType{data}
\name{tripmats}
\alias{tripmats}
\title{tripmats}
\description{
Station-to-station trip matrices between bicycle hire stations for all cities
of the \pkg{bikedata} package, calculated with the \code{bike_tripmat}
function. These matrices have been reconciled against the corresponding
\link{distmats} matrices with the function
\code{bikedata::bike_match_matrices}.
}
