% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-models.R
\name{width_ratios}
\alias{width_ratios}
\title{width_ratios}
\usage{
width_ratios(city)
}
\arguments{
\item{city}{City for which ratios are to be calculated}
}
\value{
\code{data.frame} of deciles, equivalent distances, and ratios of
both exponential decay width parameters (\code{k}), and squared errors
(\code{ss}).
}
\description{
Calculate ratios of widths of decay parameters for covariances calculated
over different distance deciles.
}
