% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-matrices.R
\name{dd_cov}
\alias{dd_cov}
\title{dd_cov}
\usage{
dd_cov(city)
}
\arguments{
\item{city}{City for which covariance matrix is to be extracted}
}
\description{
Calculate station-to-station covariance matrix for a given city. Upper
diagonal of matrix is between all trips \strong{from} i and j, so
\code{dd_cov()[i, j]} holds the covariance between all trips \strong{from}
station \code{i} to all other stations and \strong{from} station \code{j} to
all other stations. Analogously, the lower diagonal (\code{dd_cov()[j, i]})
holds covariances between all trips \strong{to} i and j.
}
\note{
The directory from which trip matrices are loaded can be retrieveed
with \link{dd_get_data_dir}, and set with \link{dd_set_data_dir}.
}
