% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov-mi-matrices.R
\name{dd_cov}
\alias{dd_cov}
\title{dd_cov}
\usage{
dd_cov(city, lower = 0, upper = 1, osm = TRUE, neutral = FALSE,
  k = NULL)
}
\arguments{
\item{city}{City for which covariance matrix is to be extracted}

\item{lower}{Lower limit (0-1) for distance cutoff used to calculate
covariances (see details)}

\item{upper}{Upper limit (0-1) for distance cutoff used to calculate
covariances (see details)}

\item{osm}{If \code{FALSE}, return straight-line distances, otherwise street
network distances.}

\item{neutral}{If \code{TRUE}, return a covariance matrix corresponding to a
neutral spatial interaction model rather than the observed trip matrix. These
neutral covariance matrices can be compared with observed ones to reveal
station pairs with higher-than-expected covariances.}

\item{k}{For neutral covariance matrices, the exponential decay parameter of
the spatial interaction model. If not specified, the value is determined by
maximising fit between neutral and observed models.}
}
\description{
Calculate station-to-station covariance matrix for a given city. Upper
diagonal of matrix is between all trips \strong{from} i and j, so
\code{dd_cov()[i, j]} holds the covariance between all trips \strong{from}
station \code{i} to all other stations and \strong{from} station \code{j} to
all other stations. Analogously, the lower diagonal (\code{dd_cov()[j, i]})
holds covariances between all trips \strong{to} i and j.
}
\note{
The directory from which trip matrices are loaded can be retrieveed
with \link{dd_get_data_dir}, and set with \link{dd_set_data_dir}. Covariances
can be calculated between stations lying within a defined distance range
using the \code{lower} and \code{upper} parameters. For example, to calculate
covariances only using the nearest half of all stations, set \code{upper =
0.5}.
}
