% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-models.R
\name{dd_decay_fns}
\alias{dd_decay_fns}
\title{dd_decay_fns}
\usage{
dd_decay_fns(city)
}
\arguments{
\item{city}{City for which decay parameters are to be calculated}

\item{from}{Plot covariance or MI values for trips \strong{from} each
station. If \code{FALSE}, plot equivalent values for trips \strong{to} each
station.}

\item{mi}{If \code{TRUE}, plot decay functions for mutual information,
otherwise plot covariances.}
}
\value{
a data frames of standardised squared residuals, p-statistics for
t-test differences between OSM and straight line versions, mean and SD
k-values quantifying the widths of exponential fits to the decay functions.
}
\description{
dd_decay_fns
}
