% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distmat.R
\name{dd_get_distmat}
\alias{dd_get_distmat}
\title{dd_get_distmat}
\usage{
dd_get_distmat(city)
}
\arguments{
\item{city}{City for which distance matrix is to be extracted}
}
\description{
Load pre-calculated distance matrix for a given city
}
\note{
The directory from which distance matrices are loaded can be retrieveed
with \link{dd_get_data_dir}, and set with \link{dd_set_data_dir}.
}
