% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_matrices.R
\name{dd_get_tripdistmats}
\alias{dd_get_tripdistmats}
\title{dd_get_tripdistsmatS}
\usage{
dd_get_tripdistmats(city, osm = TRUE)
}
\arguments{
\item{city}{City for which matrices are to be extracted}

\item{osm}{If \code{FALSE}, return straight-line distances, otherwise street
network distances.}
}
\value{
List of two square matrices of (i) numbers of trips (\code{$trip})
and (ii) distances (\code{$dist}) between all stations for nominated city.
}
\description{
Load pre-calcualted trip \strong{and} distance matrices for a given city,
ensuring that all row and column names match.
}
\note{
The directory from which trip matrices are loaded can be retrieveed
with \link{dd_get_data_dir}, and set with \link{dd_set_data_dir}.
}
