% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare-matrices.R
\name{dd_mi}
\alias{dd_mi}
\title{dd_mi}
\usage{
dd_mi(city)
}
\arguments{
\item{city}{City for which mutual information matrix is to be extracted}
}
\description{
Calculate station-to-station mutual information matrix for a given city
}
\note{
The directory from which trip matrices are loaded can be retrieveed
with \link{dd_get_data_dir}, and set with \link{dd_set_data_dir}.
}
